Moved java files into subdirectories by group
