+++ app/src/main/java/study_dev/testbed/studyhopper/Login.java
@@ -17,6 +17,7 @@ import com.google.firebase.auth.AuthResult;
 import com.google.firebase.auth.FirebaseAuth;
 import com.google.firebase.auth.FirebaseUser;
 
+import study_dev.testbed.studyhopper.ui.dashboard.Dashboard;
 import study_dev.testbed.studyhopper.ui.profile.ProfilePage;
 
 public class Login extends AppCompatActivity {
