+++ app/src/main/java/study_dev/testbed/studyhopper/StudyRoomReservations.java
@@ -1,10 +1,12 @@
 package study_dev.testbed.studyhopper;
 
+import android.content.Intent;
+import android.os.Bundle;
+
 import androidx.appcompat.app.ActionBar;
 import androidx.appcompat.app.AppCompatActivity;
 
-import android.content.Intent;
-import android.os.Bundle;
+import study_dev.testbed.studyhopper.ui.dashboard.Dashboard;
 
 public class StudyRoomReservations extends AppCompatActivity {
