--- app/src/main/java/study_dev/testbed/studyhopper/ui/dashboard/DashboardViewModel.java
@@ -1,19 +1,0 @@
-package study_dev.testbed.studyhopper.ui.dashboard;
-
-import androidx.lifecycle.LiveData;
-import androidx.lifecycle.MutableLiveData;
-import androidx.lifecycle.ViewModel;
-
-public class DashboardViewModel extends ViewModel {
-
-    private final MutableLiveData<String> mText;
-
-    public DashboardViewModel() {
-        mText = new MutableLiveData<>();
-        mText.setValue("This is dashboard fragment");
-    }
-
-    LiveData<String> getText() {
-        return mText;
-    }
-}
