+++ app/src/main/java/study_dev/testbed/studyhopper/ui/messages/Messages.java
@@ -1,4 +1,4 @@
-package study_dev.testbed.studyhopper;
+package study_dev.testbed.studyhopper.ui.messages;
 
 import android.content.Intent;
 import android.os.Bundle;
@@ -6,19 +6,27 @@ import android.os.Bundle;
 import androidx.appcompat.app.ActionBar;
 import androidx.appcompat.app.AppCompatActivity;
 
-public class StudyGroupFinder extends AppCompatActivity {
+import study_dev.testbed.studyhopper.R;
+import study_dev.testbed.studyhopper.models.studyGroupItem;
+import study_dev.testbed.studyhopper.ui.studyGroup.StudyGroupActivity;
+
+public class Messages extends AppCompatActivity {
+    private studyGroupItem item;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_study_group_finder);
+        setContentView(R.layout.activity_messages);
+
+        Intent intent = getIntent();
+        item = intent.getParcelableExtra("Study Group Data");
 
         // Enable back button
         ActionBar supportActionBar = getSupportActionBar();
         if (supportActionBar != null) {
             supportActionBar.setDisplayHomeAsUpEnabled(true);
         }
-      
+
     }
 
     @Override
@@ -29,9 +37,11 @@ public class StudyGroupFinder extends AppCompatActivity {
 
     @Override
     public void onBackPressed() {
-        Intent in = new Intent(getBaseContext(), Dashboard.class);
+
+
+        Intent in = new Intent(getBaseContext(), StudyGroupActivity.class);
+        in.putExtra("Study Group Item", item);
         startActivity(in);
         overridePendingTransition(0, 0);
-
     }
 }
