+++ app/src/main/java/study_dev/testbed/studyhopper/ui/studyLocation/StudyLocationsFragment.java
@@ -1,4 +1,4 @@
-package study_dev.testbed.studyhopper;
+package study_dev.testbed.studyhopper.ui.studyLocation;
 
 import android.os.Bundle;
 import android.view.LayoutInflater;
@@ -8,6 +8,8 @@ import android.view.ViewGroup;
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 import androidx.fragment.app.Fragment;
+
+import study_dev.testbed.studyhopper.R;
 
 public class StudyLocationsFragment extends Fragment {
