+++ app/src/main/res/layout/activity_dashboard.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:fitsSystemWindows="true"
-    tools:context=".Dashboard"
+    tools:context=".ui.dashboard.Dashboard"
     tools:openDrawer="start">
 
     <LinearLayout
