+++ app/src/main/res/layout/activity_my_groups.xml
@@ -4,8 +4,8 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".MyGroups">
-    
+    tools:context=".ui.dashboard.MyGroups">
+
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/group_recycler_view"
         android:layout_width="match_parent"
