+++ app/src/main/res/layout/activity_study_group.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".StudyGroupActivity">
+    tools:context=".ui.studyGroup.StudyGroupActivity">
 
     <com.google.android.material.floatingactionbutton.FloatingActionButton
         android:id="@+id/messageFAB"
