+++ app/src/main/res/layout/activity_study_group_finder.xml
@@ -3,6 +3,6 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".StudyGroupFinder">
+    tools:context=".ui.groupFinder.StudyGroupFinder">
 
 </androidx.constraintlayout.widget.ConstraintLayout>
